fix: add license, change name to ai_coach
