+++ ai_coach/__init__.py
@@ -1,0 +1,1 @@
+from .ai_coach import AICoachXBlock
