--- ai_tutor/__init__.py
@@ -1,1 +1,0 @@
-from .ai_tutor import AITutorXBlock
